Separate supporterProductData.ts into its own module
