+++ tsconfig.json
@@ -36,6 +36,7 @@
       "@modules/product-catalog/*": ["./modules/product-catalog/src/*"],
       "@modules/email/*": ["./modules/email/src/*"],
       "@modules/internationalisation/*": ["./modules/internationalisation/src/*"],
+      "@modules/supporter-product-data/*": ["./modules/supporter-product-data/src/*"],
       "@modules/*": ["./modules/*"]
     },
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
